CHORE/#19: 저장소 관련 파일추가, 코드수정, 패키지구조변경
