+++ app/src/main/java/com/sopt/now/compose/container/impl/AuthRepositoryImpl.kt
@@ -1,21 +1,14 @@
-package com.sopt.now.compose.container
+package com.sopt.now.compose.container.impl
 
+import com.sopt.now.compose.container.repository.AuthRepository
 import com.sopt.now.compose.network.AuthService
 import com.sopt.now.compose.network.dto.RequestLoginDto
 import com.sopt.now.compose.network.dto.RequestSignUpDto
 import com.sopt.now.compose.network.dto.ResponseLoginDto
 import com.sopt.now.compose.network.dto.ResponseSignUpDto
-import com.sopt.now.compose.ui.screens.login.LoginViewModel
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.jsonObject
 import retrofit2.Response
 
-interface AuthRepository {
-    suspend fun postLogin(request: RequestLoginDto): Result<Response<ResponseLoginDto>>
-    suspend fun postSignUp(request: RequestSignUpDto): Result<Response<ResponseSignUpDto>>
-}
-
-class NetworkAuthRepository(
+class AuthRepositoryImpl(
     private val authService: AuthService
 ) : AuthRepository {
     override suspend fun postLogin(request: RequestLoginDto): Result<Response<ResponseLoginDto>> =
