+++ app/src/main/java/com/sopt/now/compose/ui/screens/profile/ProfileViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
 import com.sopt.now.compose.SoptApplication
-import com.sopt.now.compose.container.NetworkMemberRepository
+import com.sopt.now.compose.container.impl.MemberRepositoryImpl
 import com.sopt.now.compose.models.User
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -15,7 +15,7 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 
 class ProfileViewModel(
-    private val authRepository: NetworkMemberRepository
+    private val authRepository: MemberRepositoryImpl
 ): ViewModel() {
     private val _uiState = MutableStateFlow<ProfileUiState>(ProfileUiState.Loading)
     val uiState: StateFlow<ProfileUiState> = _uiState.asStateFlow()
